Create canonical hub read model service
